requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
requests
python-dotenv
prometheus_client
